docs: Remove misleading rustc version badge
In the cargo subcommand, there are two different concepts on "supported
rustc version": "rustc version required for installation" and "rustc
version required for execution". A single badge is not sufficient to
indicate this. The "Installation" section already has a description of
this.

As for the rust-version field in Cargo.toml, it is obvious that it is
"rustc version required for installation" because of its behavior.